update changes version 0.0.7

+++ lib/models/carbon_model.dart
@@ -25,7 +25,7 @@ class CarbonModel {
   factory CarbonModel.fromJson(Map<String, dynamic> json) => CarbonModel(
         userId: json["user_id"],
         emision: json["emision"],
-        offset: json["offset"],
+        offset: json["offset"].toDouble(),
         lastCalculate: DateTime.parse(json["last_calculate"]),
       );
